Minimize exported public headers
