+++ GraphQLTree.h
@@ -3,8 +3,7 @@
 
 #pragma once
 
-#include "GraphQLGrammar.h"
-
+#include <tao/pegtl.hpp>
 #include <tao/pegtl/contrib/parse_tree.hpp>
 
 #include <string>
@@ -21,19 +20,6 @@ struct ast_node
 {
 	std::string unescaped;
 };
-
-template <typename _Rule>
-void for_each_child(const ast_node& n, std::function<bool(const ast_node&)>&& func)
-{
-	for (const auto& child : n.children)
-	{
-		if (child->is<_Rule>()
-			&& !func(*child))
-		{
-			return;
-		}
-	}
-}
 
 std::unique_ptr<ast_node> parseString(const char* text);
 std::unique_ptr<ast_node> parseFile(file_input<>&& in);
